pyproject.toml
[tool.poetry]
name = "daily-event-publisher"
version = "0.1.0"
description = ""
authors = ["Dylon Condon <dyloncondon@gmail.com>"]
readme = ".github/README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
google-api-python-client = "^2.99.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pre-commit = "^3.2.2"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
pytest-env = "^0.8.1"
pytest-cov = "^5.0.0"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
pythonpath = ["src"]
env = [
    "GOOGLE_CREDENTIALS = []",
    "GOOGLE_CALENDAR_ID = some-id",
    "TELEGRAM_API = some-api",
    "TELEGRAM_CHAT_ID = some-id",
]
